+++ ProesBack/Domain/Interfaces/INotificationRepository.cs
@@ -5,10 +5,10 @@ namespace ProesBack.Domain.Interfaces
     public interface INotificationRepository
     {
         IEnumerable<Notification> GetAll();
-        Notification Get(int id);
+        Notification Get(long id);
         void Insert(Notification notification);
         void Update(Notification notification);
-        void Delete(int id);
+        void Delete(long id);
         void Save();
     }
 }
